feat(client): move Sidebar to AppSidebar folder for reuse
