+++ packages/client/src/pages/Dashboard/Sidebar/DashboardLeftSidebar.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
+import { Siderbar } from '../../../components/AppSidebar/Sidebar';
 import { NeedsPractice } from './NeedsPractice';
 import { RecentlyLearned } from './RecentlyLearned';
 import { TrainConcepts } from './TrainConcepts';
-import { Siderbar } from './DashbaordSideBar.styles';
 
 export const DashboardLeftSidebar: React.FC<{}> = () =>
   <Siderbar>
